Calculations
To make life easier the sketch makes a number of calculations based on the hardware and optical parameters as input by the user.   These calculations include Magnification, derived from the Sensor used and the target width as measured through the viewfinder.   The Depth of Field, the Steps Between Shutters, which depends on amongst other things the microstep size selected, the actual Step size which depends on the hardware selected e.g. leadscrew pitch and stepper angle.