Run Button
The actual Run - either 'Live' with shutter actuations or 'Test' as selected in the input screen.

The Run will fire the shutter at the start of the run, the IN point and a points 'looped' after input delay(s) to allow vibrations to subside and flash-guns to recycle.   At the end of the Run the stepper will return to the IN point ready for a repeat run if called for.